% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_cox.R
\name{box_cox}
\alias{box_cox}
\title{box-cox and inverse box-cox transform}
\usage{
box_cox(x, lambda, inverse = FALSE)
}
\arguments{
\item{x}{data}

\item{lambda}{transform parameter}

\item{inverse}{logical.  Inverse transform? Defaults is FALSE (use inverse=TRUE to reverse the transform, i.e. back to natural units)}
}
\value{
a matrix of transformed data that is as close to multi-variate normal as possible with this transform; different variables are in different columns
}
\description{
Box-Cox and inverse Box-Cox power transformation (often used to normalize).
}
