% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cline.R
\name{cline}
\alias{cline}
\title{Add colored line segments to a plot}
\usage{
cline(x, y, z, color_vector)
}
\arguments{
\item{x}{x positions of points to be plotted}

\item{y}{y positions of points to be plotted}

\item{z}{a factor, the same length as x and y. Line segments in the resulting plot will be colored according to the levels of z.}

\item{color_vector}{a list of colors to use (length should match the number of levels in z).}
}
\description{
This function adds colored line segments to an existing (base-R graphics) plot.  The line is plotted at points specified by inputs x and y, and colored according to factor input z (with one color for each level of z).
}
\examples{
d <- data.frame(x=c(1:200), y=c(1:200), z=gl(4,50))
plot(d$x,d$y)
cline(d$x,d$y,d$z, color_vector=c('black', 'blue', 'red', 'orange'))
#note a ggformula alternative to using cline at all:
#gf_path(x~y, color=~z, data=d)
}
\keyword{time-series}
\keyword{visualization,}
