% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_sim.R
\name{mc_sim}
\alias{mc_sim}
\title{simulate from a Markov chain}
\usage{
mc_sim(sim_dives, n, G, G_exp, first)
}
\arguments{
\item{sim_dives}{data frame with variable "resp" with values "baseline" and "reponse"}

\item{n}{number of time-steps to simulate}

\item{G}{transition probability matrix (baseline)}

\item{G_exp}{transition probability matrix (exposure)}

\item{first}{l}
}
\value{
a vector of numeric values (1 up to number of rows/columns in G) indicating the simulated state sequence
}
\description{
simulate from a Markov chain
}
