% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sincos2angle.R
\name{sincos2angle}
\alias{sincos2angle}
\title{convert sine and cosine to angle}
\usage{
sincos2angle(sin, cos)
}
\arguments{
\item{sin}{sine of the angle}

\item{cos}{cosine of the angle}
}
\value{
the angle
}
\description{
convert sine and cosine to an angle
}
