% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_sim.R
\name{tag_sim}
\alias{tag_sim}
\title{Simulate data from an animal-borne tag}
\usage{
tag_sim(n, G, G_exp = G, first = NULL, phase_params,
  phase_params_exp = phase_params, fine_params,
  fine_params_exp = fine_params, transform_params = NULL,
  sampling_rate = 1, resp_start = 0, resp_dur = 0)
}
\arguments{
\item{n}{number of dives (or flights, although this is untested) to simulate}

\item{G}{transition probability matrix indicating probability of transition between a set of dive (or perhaps flight) types}

\item{G_exp}{transition probability matrix during experimental exposure (if any). Defaults to G (i.e., no exposure or no change in TPM during exposure).}

\item{first}{integer; dive type for the first dive.  1 corresponds to the first row of G, 2 to the second, etc. If NULL (the default), the first dive type will be a random sample from the stationary distribution determined from G.}

\item{phase_params}{parameters of multivariate normal distribution fit to (perhaps transformed) durations of different dive phases, by dive type. Assumed to be output from mclust::mvn, in terms of format.}

\item{phase_params_exp}{same as mvn_params, but during experimental exposure (if any). Defaults to mvn_est (i.e., no exposure or no change in MVN params during exposure)}

\item{fine_params}{parameters of multivariate AR(1) process fit to data for each observed tag data stream (e.g. acceleration, depth, pitch, roll...) within each phase. Assumed to be output from mAr::mAr.est, in terms of format.}

\item{fine_params_exp}{same as fine_params, but during experimental exposure (if any). Defaults to fine_params (i.e., no exposure or no change in within-phase fine-scale data during exposure).}

\item{transform_params}{parameters related to normalizing Box-Cox transformations of dive phase data (if used before estimating MVN parameters for dive phase durations).}

\item{sampling_rate}{sampling rate in Hz (samples per second) for fine-scale time-series data. Defaults to 1 Hz.}

\item{resp_start}{start time (as dive number) of simulated "response" to experimental exposure, if any. Defaults to 0 (no response).}

\item{resp_dur}{duration (in dives) of simulated "response" to experimental exposure. Defaults to 0. (NOTE - IT MAY BE OF INTEREST TO ALLOW FOR THIS and resp_start TO BE SPECIFIED IN UNITS OF EITHER DIVES, OR TIME)}
}
\value{
a data.table object containing simulated tag data. Columns in the data.table, and their names, will correspond to inputs provided to tag_sim.
}
\description{
This function generates simulated data from an animal-borne tag. It was written for DTAG data collected on whales, but may be applicable in other contexts; more changes would be required if the animal doesn't do depth/altitude excursions (dives or flights) most of the time, or if the data are not high-resolution (DTAGs collect acceleration and magnetic field and depth data 5-500 times per second).
}
