% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_sim.R
\name{mc_sim}
\alias{mc_sim}
\title{simulate from a Markov chain}
\usage{
mc_sim(n, G, G_resp, first)
}
\arguments{
\item{n}{number of time-steps to simulate}

\item{G}{transition probability matrix (baseline)}

\item{first}{l}

\item{G_exp}{transition probability matrix (exposure)}
}
\value{
a vector of numeric values (1 up to number of rows/columns in G) indicating the simulated state sequence
}
\description{
simulate from a Markov chain
}
